requirements.txt
venusian==1.0a8
cornice==0.16.2
SQLAlchemy==0.9.1
Chameleon==2.14
Jinja2==2.7.1
Mako==0.9.1
MarkupSafe==0.18
Paste==1.7.5.1
PasteDeploy==1.5.2
PasteScript==1.7.5
WebOb==1.3.1
pyramid==1.4.3
pyramid_jinja2==1.9
pyramid_debugtoolbar==1.0.9
repoze.lru==0.6
simplejson==3.3.1
translationstring==1.1
wsgi_intercept==0.6
zope.component==4.1.0
zope.deprecation==4.1.0
zope.event==4.0.2
zope.interface==4.0.5
psutil==1.2.1
ordereddict==1.1
unittest2==0.5.1
WebTest==2.0.11
Sphinx==1.2
docutils==0.11
coverage==3.7.1
greenlet==0.4.1
Pygments==1.6
cef==0.5
PyMySQL==0.6.1
pymysql_sa==1.0
protobuf==2.5
gevent==1.0
umemcache==1.6.3
requests==2.2.1
chardet==2.2.1
metlog-py==0.10
https://github.com/mozilla-services/mozservices/archive/73c13647328546fa4e906ce99f95c2fd71b51877.zip
powerhose==0.7
tokenlib==0.3.1
PyBrowserID==0.9.1
https://argparse.googlecode.com/files/argparse-1.2.1.tar.gz
circus==0.10
hawkauthlib==0.1.1
boto==2.27
